+++ packages/houdini-svelte/src/plugin/fsPatch.ts
@@ -1,7 +1,7 @@
-import type { PathLike } from 'fs'
-import filesystem, { Dirent } from 'fs'
 import type { PluginHooks } from 'houdini'
 import { fs, path } from 'houdini'
+import type { PathLike } from 'node:fs'
+import filesystem, { Dirent } from 'node:fs'
 
 import { _config } from '.'
 import type { Framework } from './kit'
@@ -176,10 +176,7 @@ filesystem.readdirSync = function (
 	}
 
 	// if there is a route component but no script, add the script
-	if (
-		contains('+page.svelte', '+page.gql') &&
-		!contains('+page.js', '+page.ts', '+page.server.js', '+page.server.ts')
-	) {
+	if (contains('+page.svelte', '+page.gql') && !contains('+page.js', '+page.ts')) {
 		result.push(virtual_file('+page.js', options))
 	}