+++ packages/houdini-svelte/src/plugin/transforms/kit/load.ts
@@ -33,7 +33,7 @@ import type { SvelteTransformPage } from '../types'
 
 const AST = recast.types.builders
 
-type ExportNamedDeclaration = ReturnType<typeof recast.types.builders['exportNamedDeclaration']>
+type ExportNamedDeclaration = ReturnType<(typeof recast.types.builders)['exportNamedDeclaration']>
 
 export default async function kit_load_generator(page: SvelteTransformPage) {
 	// if this isn't a route, move on