+++ packages/houdini/src/codegen/generators/typescript/typeReference.ts
@@ -1,9 +1,9 @@
-import type { TSTypeKind } from 'ast-types/lib/gen/kinds'
+import type { StatementKind, TSTypeKind } from 'ast-types/lib/gen/kinds'
 import * as graphql from 'graphql'
 import * as recast from 'recast'
 
+import { ensureImports, TypeWrapper, unwrapType } from '../../../lib'
 import type { Config } from '../../../lib'
-import { TypeWrapper, unwrapType } from '../../../lib'
 import { nullableField, scalarPropertyValue } from './types'
 
 const AST = recast.types.builders
@@ -18,7 +18,8 @@ export function tsTypeReference(
 			| graphql.GraphQLInputType
 			| graphql.GraphQLNamedType
 			| graphql.TypeNode
-	}
+	},
+	body: StatementKind[]
 ): TSTypeKind {
 	const { type, wrappers } = unwrapType(config, definition.type)
 
@@ -27,6 +28,10 @@ export function tsTypeReference(
 	// if we're looking at a scalar
 	if (graphql.isScalarType(type)) {
 		result = scalarPropertyValue(config, missingScalars, type)
+	}
+	//  enums need to be passed to ValueOf
+	else if (graphql.isEnumType(type)) {
+		result = enumReference(config, body, type.name)
 	}
 	// we're looking at an object
 	else {
@@ -47,4 +52,19 @@ export function tsTypeReference(
 	}
 
 	return result
+}
+
+export function enumReference(config: Config, body: StatementKind[], name: string) {
+	// if we looking at an enum we need ValueOf<enum>
+	ensureImports({
+		config,
+		body,
+		import: ['ValueOf'],
+		importKind: 'type',
+		sourceModule: '$houdini/runtime/lib/types',
+	})
+	return AST.tsTypeReference(
+		AST.identifier('ValueOf'),
+		AST.tsTypeParameterInstantiation([AST.tsTypeQuery(AST.identifier(name))])
+	)
 }