+++ packages/houdini/src/runtime/lib/types.ts
@@ -216,3 +216,7 @@ export type RequestPayload<GraphQLObject = any> = {
 		  }[]
 		| null
 }
+
+export type NestedList<_Result = string> = (_Result | null | NestedList<_Result>)[]
+
+export type ValueOf<Parent> = Parent[keyof Parent]