+++ packages/houdini-react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # houdini-react
 
+## 1.0.0-next.10
+
+### Patch Changes
+
+-   Updated dependencies [[`fd7b46c`](https://github.com/HoudiniGraphql/houdini/commit/fd7b46c4ab5392e643a6e6bb243697147d13fd2b)]:
+    -   houdini@1.0.0-next.10
+
+## 1.0.0-next.9
+
+### Patch Changes
+
+-   Updated dependencies [[`de44252`](https://github.com/HoudiniGraphql/houdini/commit/de442526e7518cc575e8f00b94767fa3d45e6f91), [`de44252`](https://github.com/HoudiniGraphql/houdini/commit/de442526e7518cc575e8f00b94767fa3d45e6f91)]:
+    -   houdini@1.0.0-next.9
+
+## 1.0.0-next.8
+
+### Dependency Changes
+
+-   Updated dependencies [[`3872a56`](https://github.com/HoudiniGraphql/houdini/commit/3872a5603b791e2530b3617bf61422e7444a483e), [`c00531f`](https://github.com/HoudiniGraphql/houdini/commit/c00531f9e5e4e57281d845816a3c92ec17faf6b8)]:
+    -   houdini@1.0.0-next.8
+
 ## 1.0.0-next.7
 
 ### Dependency Changes
@@ -31,6 +52,13 @@
 ## 1.0.0-next.3
 
 ## 1.0.0-next.2
+
+## 0.20.4
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   houdini@0.20.4
 
 ## 0.20.3